requirements.txt
chromadb==1.0.6
llama-index-core==0.12.32
llama-index-llms-ollama==0.5.4
llama-index-vector-stores-chroma==0.4.1
llama-index-embeddings-huggingface==0.5.3
llama-index-readers-file==0.4.7
ollama==0.4.8
sentence-transformers==4.1.0
torch==2.6.0
transformers==4.51.3